Added workaround to fix routes being overrided by child Controller.
When overriding a parent bundle configured with annotations routing all routes
on the parent will be ignored.

That's because if you use "@" as prefix for the controller this will pass to the kernel resolver
which will return ONLY the child resource if the parent resource has been overridden.
The solution is to provide the full path of the bundle. Considering the kernel will
try to match the resource from app/Resources, you will have to add a relative directory
(../../) before the actual path: