#10058 NewAdminBundle: Added queryBuilder to 'isUserAllowed...'
By using a query builder, even if the series is too big there won't be a memory problem thanks to the result being a Cursor.
However, the way this implemented is right now it could take enough to delete series with many, many objects to show a bad user experience. (But who would want to delete such a big series anyways?)