#213 NewAdminBundle: Fixed 'orange-box' css.
This silly 'css class through jquery' was causing a couple of bugs.
There was no need for it, so I removed the js logic and added css rules instead.
I had to change the html structure a little (I switched to using <label for=...>, then the css would work just fine. (after a minor tweak or two)