Fixed javascript error:
Since the line:
   var leftOffset = (.offset()[left]);
gives error, the rest of the code is not executed. This error is due to the id 'content' being removed on: d71fe7f67225a6776bc5cfda84efb9f93eb28272

If that is fixed, the current code creates an infinite while loop.

I have opted for removing the logic altogether.

As a note, the function it was a string shortener for when the breadcrumb title takes too much space, I believe.