TTK-17149 Use the mongo cursor to count total objects
Using Mongo cursor.count() to count the number of results for this query
and avoid an extra query.

The MongoAdapter uses cursor.count() but DoctrineODMMongoDBAdapter
doesn't use it. A new Pagerfanta Adapter is created to use this
advantage.

Force the calling getCurrentPageResults before getNbResults is needed to
keep the Pagerfanta API.

https://docs.mongodb.com/manual/reference/method/cursor.count/
https://github.com/doctrine/mongodb/blob/1.6.1/lib/Doctrine/MongoDB/Query/Query.php#L129