TTK-22509 Bugfix login attempts (#15)
* Fixed loginAttempt count: 
It's counter-intuitive that the user is not disabled until MAX_LOGIN_ATTEMPTS + 1 calls to the addLoginAttempt function. Moving the counter increment to the top of the function solves this

* Fixed weird lastLoginAttempt behaviour:
Since DateTimes are passed by reference,  is a reference to ->lastLoginAttempt, and thus
when we add the DateInterval, it is also added to lastLoginAttempt. Using 'clone' should solve this

* Reset login attempts count if interval passes
* Include interval check on User class
* Renamed functions and variables for clarity
* Using 'createFromDateString' for readability. Specially if we want to allow configuring this parameter

Co-authored-by: Pablo Ogando Ferreira <pabloogando5@gmail.com>
* fix code style
* fix code
* recover getGroups annotation to fix phpstan level 4